+++ src/main/java/org/kumoricon/registration/model/SearchSuggestion.java
@@ -6,8 +6,8 @@ import java.util.Set;
  * Represents a list of suggestions being sent back to the jQuery-Autocomplete library
  */
 public class SearchSuggestion {
-    private String query;
-    private Set<String> suggestions;
+    private final String query;
+    private final Set<String> suggestions;
 
     public SearchSuggestion(String query, Set<String> suggestions) {
         this.query = query;