+++ CHANGES.md
@@ -17,6 +17,7 @@ Version 1.0.1 (11/14/2019)
 - Staff check in: don't redirect away from step 4 when printing badge front (since they'll need to print 
   badge back as well)
 - Staff check in: added toggle to show checked in/not checked in staff
+- Add autocomplete for attendee search
 
 Version 1.0.0 (11/11/2019)
 - Renamed "Staff" report to "Users" on menu